requirements.txt
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.51
langchain_ollama==0.3.1
langchain_openai==0.3.12
langgraph==0.3.30
ollama==0.4.7
typing_extensions==4.13.2
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.7
python-dotenv==1.1.0